Dialogs/Settings: Reorganize weather configuration menu
Move weather settings out of Setup into a dedicated top-level Weather
section with separate panels for XCTherm, RASP, Flugwetter, and thermal
map. Keep user repositories on Site Files; RASP file selection moves to
the RASP panel. Add deep links for xcsoar://config/weather/* pages.
